Recover lost input focus when ringing bell
When an application steals focus from input it becomes impossible to
control Ratpoison. An example (save any work before trying this since
you most likely will have to restart Ratpoison):

1. Create prompt.html with the following content:

<html><script>setTimeout(() => { prompt(); }, 3000);</script></html>

2. Open it with Chromium (tested with Chromium 65.0.3325.146)

3. Before the prompt pops up run `exec` (C-t !)

When the prompt pops up it will steal focus from the Ratpoison
input. There does not seem to be an easy way to recover from this,
since the default top keybindings provide no way to close the input.

A more realistic case in which this scenario happens is when opening a
web page protected with HTTP basic access authentication and trying to
open your password manager before the authentication prompt pops up.

To solve this refocus the input when ringing the bell. The bell is
rung when a key is pressed with a modifier that does not result in an
action. This includes `escape` (C-t), probably one of the first
commands a user will try to resolve this.

To do this without duplicating code this moves the already existing
focusing code in a reusable helper function.