Support regular expressions in the unmanage list
CLion is a Java application and is a bit weird. Popups created by
right clicking or by using the menus at the top are being spawned as
regular windows. So ratpoison ends up managing them as a regular
window and puts them in the middle of the screen. This can be solved
with the "unmanage" feature. Putting a window name in that list will
make the window responsible for its position. This works well but for
some crazy reason each window being spawned has a counter:

win0
win1
win2
...

To solve this we'll parse every name in the unmanage list as a regular
expression. With this patch these windows can be ignored by adding

unmanage ^win[0-9]\+$

to .ratpoisorc.